1. fpgaServer.sendBlocksize(v.ChainID, block.Data.Data.len)
Before the tx signature verification in the tx validation, send the total amount of the txs in the block so that the FPGA grpc server knows when to pack the block
If we use sync.WaitGroup to send the signal of pack the txs into the hardware block, the code change made to Fabric will be very urgly (several places will need to add an extra parameter WaitGroup).

2. fpgaServer.VerifySig4VSCC(pubkeybytes_encoded, signedData_encoded, Signature_encoded)
for now, we don't support multiple channels because we can't get the channelID directly from here.
and we don't want to add this parameter (channelID) to every function call in the whole call stack. We will redesign later.

Change-Id: I556cedc6174c31f2e15f5d130d398388acf8ffaa